Alura/MyDrf/requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
Faker==15.1.1
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
validate-docbr==1.10.0
